style : password 필드 타입 수정
password의 타입을 Integer → Long 으로 수정

+++ src/main/java/com/example/scheduler/dto/ScheduleRequestDto.java
@@ -7,6 +7,6 @@ public class ScheduleRequestDto {
 
     private String title;
     private String description;
-    private Integer password;
+    private Long password;
 
 }
