+++ src/main/java/com/example/scheduler/entity/Schedule.java
@@ -13,7 +13,7 @@ public class Schedule {
     private Integer id;
     private String title;
     private String description;
-    private Integer password;
+    private Long password;
     private LocalDate startPeriod;
 
     public void updateTitle(ScheduleRequestDto dto) {
